* FIX: Fixed warnings on xrEngine (#13)

+++ code/engine.vc2008/xrEngine/FDemoPlay.cpp
@@ -55,7 +55,7 @@ CDemoPlay::CDemoPlay(const char *name, float ms, u32 cycles, float life_time) : 
 		}
 		
 		seq.resize		(sz/sizeof(Fmatrix));
-		m_count			= seq.size();
+		m_count			= (int)seq.size();
         std::memcpy(&*seq.begin(),fs->pointer(),sz);
 		FS.r_close		(fs);
 		Log				("~ Total key-frames: ",m_count);
