+++ code/engine.vc2008/xrEngine/IGame_Level.cpp
@@ -188,7 +188,7 @@ void	IGame_Level::OnFrame		( )
 		Sounds_Random_dwNextTime		= Device.dwTimeGlobal + ::Random.randI	(10000,20000);
 		Fvector	pos;
 		pos.random_dir().normalize().mul(::Random.randF(30,100)).add	(Device.vCameraPosition);
-		int		id						= ::Random.randI(Sounds_Random.size());
+		int		id						= ::Random.randI(s32(Sounds_Random.size()));
 		if (Sounds_Random_Enabled)		{
 			Sounds_Random[id].play_at_pos	(0,pos,0);
 			Sounds_Random[id].set_volume	(1.f);
