+++ code/engine.vc2008/xrEngine/StatGraph.h
@@ -127,7 +127,7 @@ public:
 	u32			AppendSubGraph	(EStyle S)
 	{
 		subgraphs.push_back(SSubGraph(S));
-		return subgraphs.size()-1;
+		return (u32)subgraphs.size()-1;
 	};
 
 	void		AddMarker (EStyle Style, float pos, u32 Color)
